requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
twilio==9.7.2
python-dotenv==1.0.0
